dom/workers/test/test_fibonacci.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads with Fibonacci
-->
<head>
  <title>Test for DOM Worker Threads with Fibonacci</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=450452">DOM Worker Threads Fibonacci</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  const seqNum = 5;

  function recursivefib(n) {
    return n < 2 ? n : recursivefib(n - 1) + recursivefib(n - 2);
  }

  var worker = new Worker("fibonacci_worker.js");

  worker.onmessage = function(event) {
    is(event.target, worker);
    is(event.data, recursivefib(seqNum));
    SimpleTest.finish();
  };

  worker.onerror = function(event) {
    is(event.target, worker);
    ok(false, "Worker had an error: " + event.data);
    SimpleTest.finish();
  };

  worker.postMessage(seqNum);

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
